リファクタリング: コードベースのクリーンアップと最適化
主な変更内容:
- 不要なファイル・ディレクトリを削除（discover-T*, logs, mbox, work, state）
- テスト用gitブランチを削除（feat/T001〜T007）
- Pythonキャッシュファイルを削除し.gitignoreを更新
- 重複テストファイルを整理・統合
- busd.pyのリファクタリング（マジックナンバー定数化、長いメソッド分割）
  - ensure_worktree: 97行→35行
  - spawn_child: 105行→30行
- busctl.pyの機能追加（環境変数サポート、--cwdパラメータ）
- テストを並列ディレクトリ方式に統一（サブディレクトリテスト削除）
- 全22個のユニットテストが成功
