+++ tests/bin/test_message_validation.py
